// Client/src/app/_models/employee.ts
export interface Employee{

        nom: string;
        prenom: string;
        cin: string;
        idSage: number;
        adresseEmail: string ;
        adresse:string;
        tel: string;
        dateNaissance: string;
        genre: string;
        situationFamiliale: string;
        dateEntree: string;
        dateSortie: string;
        soldeConge: number;
        statut: string;
        badge: boolean;
        userName: string;
        password: boolean;
        token: string;
        roleName: string ;
        refPlanningWeek: string ;
        opération:string;
        logRef:string;

      }

      
   
    
    
    
    
    
    
    