// Client/src/app/_models/planning.ts
export interface Planning {
    refPlanningWeek: string | null;
    user: string | null;
    opération: null;
    lundi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    mardi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    mercredi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    jeudi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    vendredi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    samedi: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S2: string | null;
      heureDebut_S2: string | null;
      heureFin_S1: string | null;
      heuresPlanifie: number | null;
    };
    dimanche: {
      id: number;
      refPlanning: string | null;
      heureDebut_S1: string | null;
      heureFin_S1: string | null;
      heureDebut_S2: string | null;
      heureFin_S2: string | null;
      heuresPlanifie: number | null;
    };
  }
  